create new component for jointly or inConcert control
